+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `:method` value to the `S3#presign` result indicating the HTTP verb that should be used (@janko-m)
+
 * Add ability to specify `method: :put` in `S3#presign` to generate data for PUT upload (@janko-m)
 
 * Return a `Struct` instead of a `Aws::S3::PresignedPost` object in `S3#presign` (@janko-m)